requirements.txt
# Data pipeline
requests>=2.31.0
python-dotenv>=1.0.0
Pillow>=10.3.0
tqdm>=4.66.2
 
# Embeddings (CPU-only, no GPU needed)
sentence-transformers>=3.0.0
 
# Vector database
chromadb>=0.5.0
 
# LLM and agent framework
langchain>=0.2.16
langchain-google-genai>=1.0.10
langgraph>=0.2.28
google-generativeai>=0.7.2
 
# Evaluation
ragas>=0.1.20
pandas>=2.2.0
seaborn>=0.13.0
matplotlib>=3.9.0
 
# Notebooks
jupyter>=1.0.0
ipykernel>=6.29.0
ipywidgets>=8.1.0
 
# Testing
pytest>=8.3.0
pytest-asyncio>=0.23.0
 
# Utilities
pydantic>=2.8.0